x/authz/simulation/decoder.go
package simulation

import (
	"bytes"
	"fmt"

	"github.com/Finschia/finschia-rdk/codec"
	"github.com/Finschia/finschia-rdk/types/kv"
	"github.com/Finschia/finschia-rdk/x/authz"
	"github.com/Finschia/finschia-rdk/x/authz/keeper"
)

// NewDecodeStore returns a decoder function closure that umarshals the KVPair's
// Value to the corresponding authz type.
func NewDecodeStore(cdc codec.Codec) func(kvA, kvB kv.Pair) string {
	return func(kvA, kvB kv.Pair) string {
		switch {
		case bytes.Equal(kvA.Key[:1], keeper.GrantKey):
			var grantA, grantB authz.Grant
			cdc.MustUnmarshal(kvA.Value, &grantA)
			cdc.MustUnmarshal(kvB.Value, &grantB)
			return fmt.Sprintf("%v\n%v", grantA, grantB)
		default:
			panic(fmt.Sprintf("invalid authz key %X", kvA.Key))
		}
	}
}
